New UI elements

+++ Nez.Portable/Core.cs
@@ -313,7 +313,7 @@ namespace Nez
 				TimeRuler.instance.render();
 
 			#if !FNA
-			drawCalls = graphicsDevice.Metrics.DrawCount;
+			drawCalls = (long)graphicsDevice.Metrics.DrawCount;
 			#endif
 			#endif
 		}
